+++ CHANGELOG.md
@@ -6,7 +6,20 @@ Most recent releases are shown at the top. Each release shows:
 - **Changed**: Additional parameters, changes to inputs or outputs, etc
 - **Fixed**: Bug fixes that don't change documented behaviour
 
-## 0.8.0 (2020-01-13)
+## 0.8.1 (2020-01-15)
+
+### New:
+- N/A
+
+### Changed:
+- N/A
+
+### Fixed:
+- fix to support multilabel text classification
+- `_prepare_dataset` no longer breaks when validation dataset has not been supplied
+
+
+## 0.8.0 (2020-01-14)
 
 ### New:
 - availability of a new, simplied interface to Hugging Face transformer models